requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements.txt setup.cfg
#
absl-py==1.0.0
    # via python-project-boilerplate (setup.cfg)
attrs==21.4.0
    # via python-project-boilerplate (setup.cfg)
six==1.16.0
    # via absl-py
